fix: move all images and icons to assets folder
