+++ Frontend/src/assets/images/templates/Rectangle1.png
