+++ Frontend/src/assets/images/templates/Rectangle2.png
