+++ Frontend/src/assets/images/templates/close.png
